requirements.txt
pandas
Flask
pymongo
requests
psycopg2-binary==2.9.10